+++ Zander.Domain/Entities/Team.cs
@@ -1,7 +1,7 @@
 ﻿namespace Zander.Domain.Entities {
 	public class Team {
 		public string Name { get; set; }
-		public long Color { get; set; }
+		public int Color { get; set; }
 		public short Score { get; set; }
 	}
 }
