+++ Zander.Domain/Entities/ZandronumServer.cs
@@ -1,7 +1,7 @@
 ﻿namespace Zander.Domain.Entities {
 	public class ZandronumServer : Server {
-		public override long Challenge {
-			get { return 199L; }
+		public override int Challenge {
+			get { return 199; }
 		}
 	}
 }
